store public key in the root repo directory
The public key will be used in the future to validate the signatures in
the repo, and it will also validate the private key passed into the
insert-apk command to prevent the use of a different key. This public
key isn't meant to be consumed by the app, so it could be preferable to
not include it in a sync to the static files server.