generate a file containing all AppMetadata JSONs
BulkAppMetadata represents a file where every (non-signature and
timestamp) line is a JSON string of [AppMetadata] for every app in the
repo. Intended to be used for fresh installs or force refreshes to not
require the app the make tons of network requests, as it is more
efficient to do everything in one network request.