migrate to SQLDelight with SQLite
Instead of having to go through every single plaintext file on the disk
as the repository state and relying on that for consistency, it's much
better to have a singular database keep track of the metadata. The
compromise is that the APKs and delta files are still stored on the
filesystem.