rework handling of the database
We restrict database access to a single thread. By using a Mutex from
kotlinx.coroutines, threads attempting to access the database while
another coroutine is executing on the database thread will be suspended,
which lets other coroutines have more throughput when needed.

The main insertion code has also been wrapped with a transaction and
multiple rollback callbacks to ensure that if an error happens during
insertion, all of the changes will be rolled back properly instead of
leaving the repository in an invalid or inconsistent state than what it
was before.